Extract rifle plugin from blunderbuss for blame-based reviewer selection
Split the blunderbuss plugin into two mutually exclusive plugins in a
shared review_assignment package: blunderbuss (random selection, original
behavior) and rifle (blame-based scoring). This addresses review feedback
that changing blunderbuss defaults would concentrate review load and work
against the contributor-ladder model.
