+++ pkg/plugins/review_assignment/blame.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package blunderbuss
+package review_assignment
 
 import (
 	"regexp"
